+++ mapl/GEOS.F90
@@ -3,42 +3,19 @@
 
 program geos
    use MAPL
-   use mapl_Cap_mod
+   use mapl_Cap_mod, only: MAPL_CapCreate, MAPL_CapRun
    use esmf
    implicit none
 
    integer :: status
-   type(ESMF_HConfig) :: hconfig
-   logical :: is_model_pet
+   type(MAPL_GriddedComponentDriver) :: driver
    type(ESMF_GridComp), allocatable :: servers(:)
 
-   call MAPL_Initialize(hconfig=hconfig, is_model_pet=is_model_pet, servers=servers, configFileNameFromArgNum=1, _RC)
-   call run_geos(hconfig, is_model_pet=is_model_pet, servers=servers, _RC)
+   call MAPL_Initialize(configFileNameFromArgNum=1, _RC)
+   call MAPL_CreateServers(servers, _RC)
+   call MAPL_CapCreate(driver, _RC)
+   call MAPL_RunServers(servers, _RC)
+   call MAPL_CapRun(driver, _RC)
    call MAPL_Finalize(_RC)
-
-contains
-
-#undef I_AM_MAIN
-#include "MAPL.h"
-
-   subroutine run_geos(hconfig, is_model_pet, servers, rc)
-      type(ESMF_HConfig), intent(inout) :: hconfig
-      logical, intent(in) :: is_model_pet
-      type(ESMF_GridComp), optional, intent(in) :: servers(:)
-      integer, optional, intent(out) :: rc
-
-      logical :: has_cap_hconfig
-      type(ESMF_HConfig) :: cap_hconfig
-      integer :: status
-
-      has_cap_hconfig = ESMF_HConfigIsDefined(hconfig, keystring='cap', _RC)
-      _ASSERT(has_cap_hconfig, 'No cap section found in configuration file')
-      cap_hconfig = ESMF_HConfigCreateAt(hconfig, keystring='cap', _RC)
-
-      call MAPL_run_driver(cap_hconfig, is_model_pet=is_model_pet, servers=servers, _RC)
-      call ESMF_HConfigDestroy(cap_hconfig, _RC)
-
-      _RETURN(_SUCCESS)
-   end subroutine run_geos
 
 end program geos